Moving things around somewhat resemble a feature-based single-module clean architecture.
