--- app/src/main/java/com/akucuki/videoplayersample/app/theme/Type.kt
@@ -1,51 +1,0 @@
-package com.akucuki.videoplayersample.app.theme
-
-import androidx.compose.material3.Typography
-import androidx.compose.ui.text.TextStyle
-import androidx.compose.ui.text.font.Font
-import androidx.compose.ui.text.font.FontFamily
-import androidx.compose.ui.text.font.FontWeight
-import androidx.compose.ui.unit.sp
-import com.akucuki.videoplayersample.R
-
-val PlayFontFamily = FontFamily(Font(R.font.play))
-
-// Set of Material typography styles to start with
-val Typography = Typography(
-    bodyLarge = TextStyle(
-        fontFamily = PlayFontFamily,
-        fontWeight = FontWeight.Normal,
-        fontSize = 16.sp,
-        lineHeight = 24.sp,
-        letterSpacing = 0.5.sp
-    ),
-    titleMedium = TextStyle(
-        fontFamily = PlayFontFamily,
-        fontSize = 16.sp,
-        fontWeight = FontWeight.SemiBold
-    ),
-    labelSmall = TextStyle(
-        fontFamily = PlayFontFamily,
-        fontSize = 12.sp,
-    ),
-    bodyMedium = TextStyle(
-        fontFamily = PlayFontFamily,
-        fontSize = 14.sp
-    )
-    /* Other default text styles to override
-    titleLarge = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Normal,
-        fontSize = 22.sp,
-        lineHeight = 28.sp,
-        letterSpacing = 0.sp
-    ),
-    labelSmall = TextStyle(
-        fontFamily = FontFamily.Default,
-        fontWeight = FontWeight.Medium,
-        fontSize = 11.sp,
-        lineHeight = 16.sp,
-        letterSpacing = 0.5.sp
-    )
-    */
-)
