+++ app/src/main/java/com/akucuki/videoplayersample/core/data/local/InstantConverter.kt
@@ -1,0 +1,17 @@
+package com.akucuki.videoplayersample.core.data.local
+
+import androidx.room.TypeConverter
+import java.time.Instant
+
+class InstantConverter {
+
+    @TypeConverter
+    fun fromTimestamp(value: Long?): Instant? {
+        return value?.let { Instant.ofEpochMilli(it) }
+    }
+
+    @TypeConverter
+    fun dateToTimestamp(date: Instant?): Long? {
+        return date?.toEpochMilli()
+    }
+}
