+++ app/src/main/java/com/akucuki/videoplayersample/core/data/remote/model/MoviesList.kt
@@ -1,0 +1,9 @@
+package com.akucuki.videoplayersample.core.data.remote.model
+
+import com.squareup.moshi.JsonClass
+
+@JsonClass(generateAdapter = true)
+data class MoviesList(
+    val name: String,
+    val videos: List<MovieData>
+)
