--- app/src/main/java/com/akucuki/videoplayersample/data/local/model/LocalVideoFetchInfo.kt
@@ -1,18 +1,0 @@
-package com.akucuki.videoplayersample.data.local.model
-
-import androidx.room.ColumnInfo
-import androidx.room.Entity
-import androidx.room.PrimaryKey
-import androidx.room.TypeConverters
-import com.akucuki.videoplayersample.data.local.InstantConverter
-import java.time.Instant
-
-@Entity(tableName = "videos_fetch_info_table")
-data class LocalVideoFetchInfo(
-    @PrimaryKey
-    @ColumnInfo("id")
-    val id: Int = 1,
-    @ColumnInfo("lastFetchDate")
-    @TypeConverters(InstantConverter::class)
-    val lastFetchDate: Instant? = null
-)
