--- app/src/main/java/com/akucuki/videoplayersample/data/remote/model/MoviesList.kt
@@ -1,9 +1,0 @@
-package com.akucuki.videoplayersample.data.remote.model
-
-import com.squareup.moshi.JsonClass
-
-@JsonClass(generateAdapter = true)
-data class MoviesList(
-    val name: String,
-    val videos: List<MovieData>
-)
