--- app/src/main/java/com/akucuki/videoplayersample/ui/screens/home/HomeEvents.kt
@@ -1,7 +1,0 @@
-package com.akucuki.videoplayersample.ui.screens.home
-
-sealed class HomeEvents {
-
-    data class ShowToast(val message: String) : HomeEvents()
-    data class NavigateToPlayer(val id: Int) : HomeEvents()
-}
