--- app/src/main/java/com/akucuki/videoplayersample/ui/screens/player/PlayerEvents.kt
@@ -1,6 +1,0 @@
-package com.akucuki.videoplayersample.ui.screens.player
-
-sealed class PlayerEvents {
-
-    data class ShowToast(val message: String) : PlayerEvents()
-}
