+++ app/src/main/java/com/akucuki/videoplayersample/videoPlayerFeature/presentation/player/PlayerEvents.kt
@@ -1,0 +1,6 @@
+package com.akucuki.videoplayersample.videoPlayerFeature.presentation.player
+
+sealed class PlayerEvents {
+
+    data class ShowToast(val message: String) : PlayerEvents()
+}
