+++ app/src/main/java/com/akucuki/videoplayersample/videoPlayerFeature/presentation/videoList/VideoListState.kt
@@ -1,0 +1,12 @@
+package com.akucuki.videoplayersample.videoPlayerFeature.presentation.videoList
+
+import android.os.Parcelable
+import androidx.compose.runtime.Immutable
+import kotlinx.parcelize.Parcelize
+
+@Immutable
+@Parcelize
+data class VideoListState(
+    val isLoading: Boolean = false,
+    val itemsData: List<VideoThumbnailUi> = emptyList()
+) : Parcelable
